refactor(client): extract global mpe integ test method in mpe test utils
